+++ packages/aim_core/lib/aim_core.dart
@@ -8,7 +8,7 @@ library;
 export 'src/app.dart';
 export 'src/body.dart';
 export 'src/context.dart';
-export 'src/env.dart';
+export 'src/variables.dart';
 export 'src/message.dart';
 export 'src/request.dart';
 export 'src/response.dart';